requirements.txt
scipy
matplotlib
parmap
DataSpyre
